single element in tcpinfo.go
