+++ pbparser/tcpinfo.go
@@ -8,8 +8,11 @@ import (
 	"encoding/binary"
 	"errors"
 	"io"
+	"log"
+	"strings"
 	"time"
 
+	"cloud.google.com/go/bigquery"
 	"github.com/gogo/protobuf/proto"
 	"github.com/prometheus/client_golang/prometheus"
 
@@ -119,4 +122,83 @@ func ReadAll(rdr io.Reader) ([]tcp.TCPDiagnosticsProto, error) {
 		inserter.PutAsync(rowBuf.TakeRows()) */
 
 	return result, nil
+}
+
+type TCPDiagnosticsProto struct {
+	InetDiagMsg *tcp.InetDiagMsgProto `protobuf:"bytes,1,opt,name=inet_diag_msg,json=inetDiagMsg,proto3" json:"inet_diag_msg,omitempty"`
+	// From INET_DIAG_PROTOCOL message.
+	DiagProtocol tcp.Protocol `protobuf:"varint,2,opt,name=diag_protocol,json=diagProtocol,proto3,enum=Protocol" json:"diag_protocol,omitempty"`
+	// From INET_DIAG_CONG message.
+	CongestionAlgorithm string `protobuf:"bytes,3,opt,name=congestion_algorithm,json=congestionAlgorithm,proto3" json:"congestion_algorithm,omitempty"`
+	// The following three are mutually exclusive, as they provide
+	// data from different congestion control strategies.
+	//
+	// Types that are valid to be assigned to CcInfo:
+	//	*TCPDiagnosticsProto_Vegas
+	//	*TCPDiagnosticsProto_Dctcp
+	//	*TCPDiagnosticsProto_BbrInfo
+	//CcInfo isTCPDiagnosticsProto_CcInfo `protobuf_oneof:"cc_info"`
+	// Data obtained from INET_DIAG_SKMEMINFO.
+	SocketMem *tcp.SocketMemInfoProto `protobuf:"bytes,7,opt,name=socket_mem,json=socketMem,proto3" json:"socket_mem,omitempty"`
+	// Data obtained from INET_DIAG_MEMINFO.
+	MemInfo *tcp.MemInfoProto `protobuf:"bytes,8,opt,name=mem_info,json=memInfo,proto3" json:"mem_info,omitempty"`
+	// Data obtained from struct tcp_info.
+	TcpInfo *tcp.TCPInfoProto `protobuf:"bytes,9,opt,name=tcp_info,json=tcpInfo,proto3" json:"tcp_info,omitempty"`
+	// If there is shutdown info, this is the mask value.
+	// Check has_shutdown_mask to determine whether present.
+	//
+	// Types that are valid to be assigned to Shutdown:
+	//	*TCPDiagnosticsProto_ShutdownMask
+	// Shutdown isTCPDiagnosticsProto_Shutdown `protobuf_oneof:"shutdown"`
+	ShutdownMask uint32
+	// Timestamp of batch of messages containing this message.
+	Timestamp time.Time `protobuf:"varint,11,opt,name=timestamp,proto3" json:"timestamp,omitempty"`
+}
+
+// InfoWrapper that implements ValueSaver
+type InfoWrapper struct {
+	tcp.TCPDiagnosticsProto
+}
+
+// Save implements the ValueSaver.Save() method.
+func (iw InfoWrapper) Save() (row map[string]bigquery.Value, insertID string, err error) {
+	// Assemble the full map by examining each top level field.
+	row = make(map[string]bigquery.Value, 10)
+	var schema bigquery.Schema
+	if iw.InetDiagMsg != nil {
+		schema, err = bigquery.InferSchema(*iw.InetDiagMsg)
+		if err != nil {
+			return
+		}
+		schema = removeXXX(schema)
+		ss := bigquery.StructSaver{Schema: schema, InsertID: "", Struct: *iw.InetDiagMsg}
+		row["InetDiagMsg"], _, err = ss.Save()
+		if err != nil {
+			return
+		}
+	}
+
+	return
+}
+
+func removeXXX(schema bigquery.Schema) bigquery.Schema {
+	result := make([]*bigquery.FieldSchema, 0, len(schema))
+	for i := range schema {
+		if !strings.HasPrefix(schema[i].Name, "XXX_") {
+			log.Println(i, schema[i].Name, schema[i].Type)
+			if schema[i].Type == bigquery.RecordFieldType {
+				schema[i].Schema = removeXXX(schema[i].Schema)
+			}
+			result = append(result, schema[i])
+		}
+	}
+	return result
+}
+
+func GetSchema() (bigquery.Schema, error) {
+	schema, err := bigquery.InferSchema(TCPDiagnosticsProto{})
+	if err != nil {
+		return bigquery.Schema{}, err
+	}
+	return removeXXX(schema), nil
 }
